+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Upcoming
+* Increased by 0.25 seconds the time before the Acid Fireball cloud starts dealing damage
+* Reduced "increased chance" and range to charge if the player is near a crystal (when 10 -> 8 blocks from a crystal) (x2 chance to charge -> x1.75 chance to charge)
+* Fixed Dragon Minion shooting twice as fast
+
 ## Beta 3.5.1
 * Wither
   * Attack